+++ clients/include/rocsolver.hpp
@@ -3312,12 +3312,11 @@ inline rocblas_status rocsolver_sygsx_hegsx(bool STRIDED,
 {
     if(STRIDED)
         return SYGST
-            ? rocblas_status_not_implemented //rocsolver_ssygst_strided_batched(handle, itype, uplo, n, A, lda, stA, B, ldb, stB, bc)
+            ? rocsolver_ssygst_strided_batched(handle, itype, uplo, n, A, lda, stA, B, ldb, stB, bc)
             : rocsolver_ssygs2_strided_batched(handle, itype, uplo, n, A, lda, stA, B, ldb, stB, bc);
     else
-        return SYGST
-            ? rocblas_status_not_implemented //rocsolver_ssygst(handle, itype, uplo, n, A, lda, B, ldb)
-            : rocsolver_ssygs2(handle, itype, uplo, n, A, lda, B, ldb);
+        return SYGST ? rocsolver_ssygst(handle, itype, uplo, n, A, lda, B, ldb)
+                     : rocsolver_ssygs2(handle, itype, uplo, n, A, lda, B, ldb);
 }
 
 inline rocblas_status rocsolver_sygsx_hegsx(bool STRIDED,
@@ -3336,12 +3335,11 @@ inline rocblas_status rocsolver_sygsx_hegsx(bool STRIDED,
 {
     if(STRIDED)
         return SYGST
-            ? rocblas_status_not_implemented //rocsolver_dsygst_strided_batched(handle, itype, uplo, n, A, lda, stA, B, ldb, stB, bc)
+            ? rocsolver_dsygst_strided_batched(handle, itype, uplo, n, A, lda, stA, B, ldb, stB, bc)
             : rocsolver_dsygs2_strided_batched(handle, itype, uplo, n, A, lda, stA, B, ldb, stB, bc);
     else
-        return SYGST
-            ? rocblas_status_not_implemented //rocsolver_dsygst(handle, itype, uplo, n, A, lda, B, ldb)
-            : rocsolver_dsygs2(handle, itype, uplo, n, A, lda, B, ldb);
+        return SYGST ? rocsolver_dsygst(handle, itype, uplo, n, A, lda, B, ldb)
+                     : rocsolver_dsygs2(handle, itype, uplo, n, A, lda, B, ldb);
 }
 
 inline rocblas_status rocsolver_sygsx_hegsx(bool STRIDED,
@@ -3360,12 +3358,11 @@ inline rocblas_status rocsolver_sygsx_hegsx(bool STRIDED,
 {
     if(STRIDED)
         return SYGST
-            ? rocblas_status_not_implemented //rocsolver_chegst_strided_batched(handle, itype, uplo, n, A, lda, stA, B, ldb, stB, bc)
+            ? rocsolver_chegst_strided_batched(handle, itype, uplo, n, A, lda, stA, B, ldb, stB, bc)
             : rocsolver_chegs2_strided_batched(handle, itype, uplo, n, A, lda, stA, B, ldb, stB, bc);
     else
-        return SYGST
-            ? rocblas_status_not_implemented //rocsolver_chegst(handle, itype, uplo, n, A, lda, B, ldb)
-            : rocsolver_chegs2(handle, itype, uplo, n, A, lda, B, ldb);
+        return SYGST ? rocsolver_chegst(handle, itype, uplo, n, A, lda, B, ldb)
+                     : rocsolver_chegs2(handle, itype, uplo, n, A, lda, B, ldb);
 }
 
 inline rocblas_status rocsolver_sygsx_hegsx(bool STRIDED,
@@ -3384,12 +3381,11 @@ inline rocblas_status rocsolver_sygsx_hegsx(bool STRIDED,
 {
     if(STRIDED)
         return SYGST
-            ? rocblas_status_not_implemented //rocsolver_zhegst_strided_batched(handle, itype, uplo, n, A, lda, stA, B, ldb, stB, bc)
+            ? rocsolver_zhegst_strided_batched(handle, itype, uplo, n, A, lda, stA, B, ldb, stB, bc)
             : rocsolver_zhegs2_strided_batched(handle, itype, uplo, n, A, lda, stA, B, ldb, stB, bc);
     else
-        return SYGST
-            ? rocblas_status_not_implemented //rocsolver_zhegst(handle, itype, uplo, n, A, lda, B, ldb)
-            : rocsolver_zhegs2(handle, itype, uplo, n, A, lda, B, ldb);
+        return SYGST ? rocsolver_zhegst(handle, itype, uplo, n, A, lda, B, ldb)
+                     : rocsolver_zhegs2(handle, itype, uplo, n, A, lda, B, ldb);
 }
 
 // batched
@@ -3407,9 +3403,8 @@ inline rocblas_status rocsolver_sygsx_hegsx(bool STRIDED,
                                             rocblas_stride stB,
                                             rocblas_int bc)
 {
-    return SYGST
-        ? rocblas_status_not_implemented //rocsolver_ssygst_batched(handle, itype, uplo, n, A, lda, B, ldb, bc)
-        : rocsolver_ssygs2_batched(handle, itype, uplo, n, A, lda, B, ldb, bc);
+    return SYGST ? rocsolver_ssygst_batched(handle, itype, uplo, n, A, lda, B, ldb, bc)
+                 : rocsolver_ssygs2_batched(handle, itype, uplo, n, A, lda, B, ldb, bc);
 }
 
 inline rocblas_status rocsolver_sygsx_hegsx(bool STRIDED,
@@ -3426,9 +3421,8 @@ inline rocblas_status rocsolver_sygsx_hegsx(bool STRIDED,
                                             rocblas_stride stB,
                                             rocblas_int bc)
 {
-    return SYGST
-        ? rocblas_status_not_implemented //rocsolver_dsygst_batched(handle, itype, uplo, n, A, lda, B, ldb, bc)
-        : rocsolver_dsygs2_batched(handle, itype, uplo, n, A, lda, B, ldb, bc);
+    return SYGST ? rocsolver_dsygst_batched(handle, itype, uplo, n, A, lda, B, ldb, bc)
+                 : rocsolver_dsygs2_batched(handle, itype, uplo, n, A, lda, B, ldb, bc);
 }
 
 inline rocblas_status rocsolver_sygsx_hegsx(bool STRIDED,
@@ -3445,9 +3439,8 @@ inline rocblas_status rocsolver_sygsx_hegsx(bool STRIDED,
                                             rocblas_stride stB,
                                             rocblas_int bc)
 {
-    return SYGST
-        ? rocblas_status_not_implemented //rocsolver_chegst_batched(handle, itype, uplo, n, A, lda, B, ldb, bc)
-        : rocsolver_chegs2_batched(handle, itype, uplo, n, A, lda, B, ldb, bc);
+    return SYGST ? rocsolver_chegst_batched(handle, itype, uplo, n, A, lda, B, ldb, bc)
+                 : rocsolver_chegs2_batched(handle, itype, uplo, n, A, lda, B, ldb, bc);
 }
 
 inline rocblas_status rocsolver_sygsx_hegsx(bool STRIDED,
@@ -3464,8 +3457,7 @@ inline rocblas_status rocsolver_sygsx_hegsx(bool STRIDED,
                                             rocblas_stride stB,
                                             rocblas_int bc)
 {
-    return SYGST
-        ? rocblas_status_not_implemented //rocsolver_zhegst_batched(handle, itype, uplo, n, A, lda, B, ldb, bc)
-        : rocsolver_zhegs2_batched(handle, itype, uplo, n, A, lda, B, ldb, bc);
+    return SYGST ? rocsolver_zhegst_batched(handle, itype, uplo, n, A, lda, B, ldb, bc)
+                 : rocsolver_zhegs2_batched(handle, itype, uplo, n, A, lda, B, ldb, bc);
 }
 /********************************************************/